Only speedy load with speedloaders
Adds a new variable to ammo boxes; multigive.
Speedloading requires the receiver to have multiload, and ammo giver to have multigive.
Adds multigive to all revolver speedloaders and the n7.62 stripper clip.

Makes the ammo loading bar appear above the ammo box that is being loaded from, so it is visible on firearms with internal magazines.

Makes the failure to load message for internal magazines more vague, so it doesn't say there is no room for a cartridge when your ammo box you're loading from is empty.

Makes the ammo loading sounds slightly quieter.